requirements-prod.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
